requirements.txt
python-dotenv==0.19.1
requests==2.26.0
opencv-python==4.5.3.56
pyautogui==0.9.53
psutil==5.8.0